refactor(backend): split promotion postgres tests
