+++ apps/backend/scripts/postgresIntegrations/boundaries.mjs
@@ -86,7 +86,9 @@ export const boundaryDefinitions = Object.freeze([
       "src/catalog/authoring/lockOrder.postgres.integration.ts",
       "src/cards/generatedImageAppend.postgres.integration.ts",
       "src/chat/cardImages/operation.postgres.integration.ts",
-      "src/chat/cardImages/promotion/jobs.postgres.integration.ts",
+      "src/chat/cardImages/promotion/jobsLeasing.postgres.integration.ts",
+      "src/chat/cardImages/promotion/jobsSettlement.postgres.integration.ts",
+      "src/chat/cardImages/promotion/jobsRevocation.postgres.integration.ts",
       "src/chat/runs/generatedImageAttemptBudget.postgres.integration.ts",
       "src/database/aiChatInitiatingAuthClassification.postgres.integration.ts",
       "src/mediaAssets/blobLifecycle/cleanup.postgres.integration.ts",
